+++ src/export.ts
@@ -1,2 +1,3 @@
 ﻿export { parse } from './renderer/parser';
 export { bind } from './renderer/bind';
+export { render } from './renderer/render';