+++ src/parser/inline/media.test.ts
@@ -11,19 +11,19 @@ describe('Unit: parser/media', () => {
       assert.deepStrictEqual(inspect(parser('![](vbscript:alert)')), void 0);
       assert.deepStrictEqual(inspect(parser('![](data:text/html;base64,PHNjcmlwdD5hbGVydCgnWFNTJyk8L3NjcmlwdD4K)')), void 0);
       assert.deepStrictEqual(inspect(parser('![](any:alert)')), void 0);
-      assert.deepStrictEqual(inspect(parser('![](#")')), [['<img class="media" data-src="#&quot;" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](#")')), [['<img class="media" data-src="#&quot;" alt="" style="max-width: 100%;">'], '']);
       assert.deepStrictEqual(inspect(parser('![](#\\)')), void 0);
-      assert.deepStrictEqual(inspect(parser('![](")')), [['<img class="media" data-src="&quot;" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![]("#)')), [['<img class="media" data-src="&quot;#" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](")')), [['<img class="media" data-src="&quot;" alt="" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![]("#)')), [['<img class="media" data-src="&quot;#" alt="" style="max-width: 100%;">'], '']);
       assert.deepStrictEqual(inspect(parser('![]("\\)')), void 0);
       assert.deepStrictEqual(inspect(parser('![](\\)')), void 0);
-      assert.deepStrictEqual(inspect(parser('![](\\#)')), [['<img class="media" data-src="#" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](\\")')), [['<img class="media" data-src="&quot;" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](\\\\)')), [['<img class="media" data-src="\\" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](\\\\#)')), [['<img class="media" data-src="\\#" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![](\\\\")')), [['<img class="media" data-src="\\&quot;" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('!["](#)')), [['<img class="media" data-src="#" alt="&quot;">'], '']);
-      assert.deepStrictEqual(inspect(parser('![\\"](#)')), [['<img class="media" data-src="#" alt="&quot;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](\\#)')), [['<img class="media" data-src="#" alt="" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](\\")')), [['<img class="media" data-src="&quot;" alt="" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](\\\\)')), [['<img class="media" data-src="\\" alt="" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](\\\\#)')), [['<img class="media" data-src="\\#" alt="" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](\\\\")')), [['<img class="media" data-src="\\&quot;" alt="" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('!["](#)')), [['<img class="media" data-src="#" alt="&quot;" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![\\"](#)')), [['<img class="media" data-src="#" alt="&quot;" style="max-width: 100%;">'], '']);
     });
 
     it('invalid', () => {
@@ -42,21 +42,21 @@ describe('Unit: parser/media', () => {
     });
 
     it('ab', () => {
-      assert.deepStrictEqual(inspect(parser('![](b)')), [['<img class="media" data-src="b" alt="">'], '']);
-      assert.deepStrictEqual(inspect(parser('![a](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
-      assert.deepStrictEqual(inspect(parser('![a b](c)')), [['<img class="media" data-src="c" alt="a b">'], '']);
-      assert.deepStrictEqual(inspect(parser('![ a](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
-      assert.deepStrictEqual(inspect(parser('![a ](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
-      assert.deepStrictEqual(inspect(parser('![ a ](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](b)')), [['<img class="media" data-src="b" alt="" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![a](b)')), [['<img class="media" data-src="b" alt="a" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![a b](c)')), [['<img class="media" data-src="c" alt="a b" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![ a](b)')), [['<img class="media" data-src="b" alt="a" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![a ](b)')), [['<img class="media" data-src="b" alt="a" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![ a ](b)')), [['<img class="media" data-src="b" alt="a" style="max-width: 100%;">'], '']);
     });
 
     it('nest', () => {
-      assert.deepStrictEqual(inspect(parser('![\\[](#)')), [['<img class="media" data-src="#" alt="[">'], '']);
-      assert.deepStrictEqual(inspect(parser('![<wbr>"]("?"#")')), [['<img class="media" data-src="&quot;?&quot;#&quot;" alt="<wbr>&quot;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![\\[](#)')), [['<img class="media" data-src="#" alt="[" style="max-width: 100%;">'], '']);
+      assert.deepStrictEqual(inspect(parser('![<wbr>"]("?"#")')), [['<img class="media" data-src="&quot;?&quot;#&quot;" alt="<wbr>&quot;" style="max-width: 100%;">'], '']);
     });
 
     it('external', () => {
-      assert.deepStrictEqual(inspect(parser('![](//example.com)')), [['<img class="media" data-src="//example.com" alt="">'], '']);
+      assert.deepStrictEqual(inspect(parser('![](//example.com)')), [['<img class="media" data-src="//example.com" alt="" style="max-width: 100%;">'], '']);
     });
 
   });