+++ src/renderer/media/pdf.ts
@@ -1,0 +1,29 @@
+﻿import DOM from 'typed-dom';
+import { parse } from '../parser';
+
+export function pdf(source: HTMLImageElement): HTMLElement | void {
+  const url = source.getAttribute('data-src')!;
+  if (!url.split(/[?#]/).shift()!.endsWith('.pdf') || url.split('/').length < 4) return;
+  return DOM.div({
+    class: 'media',
+    style: 'position: relative;',
+  }, [
+    DOM.div({
+      style: 'position: relative; resize: vertical; overflow: hidden; padding-bottom: 10px;',
+    }, [
+      DOM.object({
+        type: 'application/pdf',
+        data: url,
+        style: 'width: 100%; height: 100%; min-height: 400px;',
+      }, [], () => {
+        const el = document.createElement('object');
+        el.typeMustMatch = true;
+        return el;
+      }),
+    ]),
+    DOM.div([
+      DOM.strong([], () =>
+        parse(`**${url}**`).querySelector('strong')!),
+    ]),
+  ]).element;
+}